Stop mixing single and double tap detection.
This changes the API to have single boolean `tapped` property
(formerly read_click).

This change aligns the behavior with the expectation that a tap
thats part of a double tap does not register as a single tap. It
will now only detect single or double taps as set by `set_tap`.

Names with click were renamed to tap so they align with what is
actually happening rather than the data sheet names. A comment in
the code points this out.

The tap example now also deduplicates taps because tap detection
holds the status for a while.